+++ scripts/specificClasses/necromancer.js
@@ -94,6 +94,51 @@ export function setNecromancerHooks() {
         }
       }
     }
+
+    // Blood Healing
+    if (
+      actor?.rollOptions?.all?.["feature:blood"] &&
+      !message?.flags?.pf2e?.context?.type
+    ) {
+      const item = message?.item;
+      if (DESTROY_THRALL_SLUGS.has(item?.slug)) {
+        const isDestroy = !["osteo-armaments", "thrall-charge"].includes(
+          item?.slug,
+        )
+          ? true
+          : await foundry.applications.api.DialogV2.confirm({
+              window: {
+                title: "pf2e-summons-assistant.dialog.thrall-destroyed.title",
+                icon: "fa-duotone fa-solid fa-person-burst",
+              },
+              content: game.i18n.localize(
+                "pf2e-summons-assistant.dialog.thrall-destroyed.content",
+              ),
+            });
+
+        if (isDestroy) {
+          const healingAmount = Math.ceil((actor?.level ?? 1) / 4);
+          const DamageRoll = CONFIG.Dice.rolls.find(
+            (r) => r.name === "DamageRoll",
+          );
+          const roll = new DamageRoll(`${healingAmount}[healing]`);
+          roll.toMessage({
+            flavor: `Blood`,
+            speaker: ChatMessage.getSpeaker({ actor }),
+            flags: {
+              "pf2e-toolbelt": {
+                targetHelper: {
+                  type: "action",
+                  author: message?.token?.document?.uuid,
+                  item: "Compendium.pf2e.classfeatures.Item.gyN8OZZ3txxIAKLf",
+                  targets: message?.token?.document?.uuid,
+                },
+              },
+            },
+          });
+        }
+      }
+    }
   });
 
   if (game.settings.get(MODULE_ID, "necromancer.thrall.auto-expire")) {